add failed buy transactions... separate logic into separate components
